+++ src/components/EnlargedPreview/index.js
@@ -1,11 +1,9 @@
 // @flow
-import * as React from 'react'
 import { h, Component } from 'preact'
 import classNames from 'classnames'
 import Pannable from '../Pannable'
 import Button from '../Button'
-import { localised } from '../../locales'
-import type { LocalisedType } from '../../locales'
+import { localised, type LocalisedType } from '../../locales'
 import {
   withNavigationDisabledState,
   withNavigationDisableAction,
@@ -110,8 +108,8 @@ class EnlargedPreview extends Component<Props, State> {
           onClick={this.toggle}
         >
           {isExpanded
-            ? translate('confirm.enlarge_image.close')
-            : translate('confirm.enlarge_image.enlarge')}
+            ? translate('doc_confirmation.button_close')
+            : translate('doc_confirmation.button_zoom')}
         </Button>
       </div>
     )