+++ src/components/Router/StepComponentMap.js
@@ -1,6 +1,6 @@
-import { h } from 'preact'
 import Welcome from '../Welcome'
 import { SelectPoADocument, SelectIdentityDocument } from '../Select'
+import CountrySelector from '../CountrySelector'
 import ImageQualityGuide from '../Uploader/ImageQualityGuide'
 import SelfieIntro from '../Photo/SelfieIntro'
 import {
@@ -22,7 +22,8 @@ import ClientSuccess from '../crossDevice/ClientSuccess'
 import CrossDeviceIntro from '../crossDevice/Intro'
 import VideoIntro from '../Video/Intro'
 import { PoACapture, PoAIntro, PoAGuidance } from '../ProofOfAddress'
-import { isDesktop, isHybrid } from '~utils'
+import { isDesktop, isHybrid, hasOnePreselectedDocument } from '~utils'
+import { getCountryDataForDocumentType } from '../../supported-documents'
 
 export const componentsList = ({
   flow,
@@ -60,8 +61,6 @@ const shouldUseVideo = (steps) => {
   )
 }
 
-const hasPreselectedDocument = (steps) => enabledDocuments(steps).length === 1
-
 const shouldUseCameraForDocumentCapture = (steps, deviceHasCameraSupport) => {
   const { options: documentOptions } = steps.find(
     (step) => step.type === 'document'
@@ -74,21 +73,20 @@ const shouldUseCameraForDocumentCapture = (steps, deviceHasCameraSupport) => {
   )
 }
 
-// This logic should not live here.
-// It should be exported into a helper when the documentType logic and routing is refactored
-export const enabledDocuments = (steps) => {
-  const documentStep = steps.find((step) => step.type === 'document')
-  const docTypes =
-    documentStep && documentStep.options && documentStep.options.documentTypes
-  return docTypes ? Object.keys(docTypes).filter((type) => docTypes[type]) : []
-}
-
 const captureStepsComponents = (
   documentType,
   mobileFlow,
   steps,
   deviceHasCameraSupport
 ) => {
+  const documentStep = steps.find((step) => step.type === 'document')
+  const documentStepOptions = documentStep?.options
+
+  // DEPRECATED: documentStep.options.showCountrySelection will be deprecated in a future release
+  const showCountrySelectionForSinglePreselectedDocument =
+    documentStepOptions?.showCountrySelection
+
+  const configForDocumentType = documentStepOptions?.documentTypes[documentType]
   const complete = mobileFlow ? [ClientSuccess] : [Complete]
   return {
     welcome: () => [Welcome],
@@ -96,8 +94,10 @@ const captureStepsComponents = (
     document: () =>
       getIdentityDocumentComponents(
         documentType,
-        hasPreselectedDocument(steps),
-        shouldUseCameraForDocumentCapture(steps, deviceHasCameraSupport)
+        hasOnePreselectedDocument(steps),
+        showCountrySelectionForSinglePreselectedDocument,
+        shouldUseCameraForDocumentCapture(steps, deviceHasCameraSupport),
+        configForDocumentType
       ),
     poa: () => [
       PoAIntro,
@@ -139,22 +139,66 @@ const getRequiredSelfieSteps = (deviceHasCameraSupport) => {
   return allSelfieSteps
 }
 
+const getNonPassportFrontDocumentCaptureFlow = (
+  hasOnePreselectedDocument,
+  showCountrySelection
+) => {
+  const frontCaptureComponents = [FrontDocumentCapture, DocumentFrontConfirm]
+  if (hasOnePreselectedDocument && showCountrySelection) {
+    return [CountrySelector, ...frontCaptureComponents]
+  }
+  if (hasOnePreselectedDocument && !showCountrySelection) {
+    return frontCaptureComponents
+  }
+  if (!hasOnePreselectedDocument && !showCountrySelection) {
+    return [SelectIdentityDocument, ...frontCaptureComponents]
+  }
+  return [SelectIdentityDocument, CountrySelector, ...frontCaptureComponents]
+}
+
 const getIdentityDocumentComponents = (
   documentType,
-  hasPreselectedDocument,
-  shouldUseCameraForDocumentCapture
+  hasOnePreselectedDocument,
+  showCountrySelectionForSinglePreselectedDocument,
+  shouldUseCameraForDocumentCapture,
+  configForDocumentType
 ) => {
-  const double_sided_docs = ['driving_licence', 'national_identity_card']
+  const doubleSidedDocs = [
+    'driving_licence',
+    'national_identity_card',
+    'residence_permit',
+  ]
+  const isPassportDocument = documentType === 'passport'
   const isDocumentUpload = !shouldUseCameraForDocumentCapture
-  const frontCaptureComponents =
-    documentType === 'passport' && isDocumentUpload
-      ? [FrontDocumentCapture, ImageQualityGuide, DocumentFrontConfirm]
-      : [FrontDocumentCapture, DocumentFrontConfirm]
-  const withSelectScreen = [SelectIdentityDocument, ...frontCaptureComponents]
-  const frontDocumentFlow = hasPreselectedDocument
-    ? frontCaptureComponents
-    : withSelectScreen
-  if (double_sided_docs.includes(documentType)) {
+
+  if (isPassportDocument) {
+    let frontCaptureComponents = [FrontDocumentCapture, DocumentFrontConfirm]
+    if (isDocumentUpload) {
+      frontCaptureComponents = [
+        FrontDocumentCapture,
+        ImageQualityGuide,
+        DocumentFrontConfirm,
+      ]
+    }
+    if (hasOnePreselectedDocument) {
+      return frontCaptureComponents
+    }
+    return [SelectIdentityDocument, ...frontCaptureComponents]
+  }
+
+  const countryCode = configForDocumentType?.country
+  const supportedCountry = getCountryDataForDocumentType(
+    countryCode,
+    documentType
+  )
+  const showCountrySelection =
+    showCountrySelectionForSinglePreselectedDocument ||
+    (!hasOnePreselectedDocument && !supportedCountry && countryCode !== null)
+  const frontDocumentFlow = getNonPassportFrontDocumentCaptureFlow(
+    hasOnePreselectedDocument,
+    showCountrySelection
+  )
+  if (doubleSidedDocs.includes(documentType)) {
     return [...frontDocumentFlow, BackDocumentCapture, DocumentBackConfirm]
   }
   return frontDocumentFlow
@@ -180,7 +224,7 @@ const createComponentList = (components, steps) => {
 const createComponent = (components, step, stepIndex) => {
   const { type } = step
   if (!(type in components)) {
-    console.error('No such step: ' + type)
+    console.error(`No such step: ${type}`)
   }
   return components[type]().map(wrapComponent(step, stepIndex))
 }