+++ test/main.js
@@ -21,9 +21,11 @@ const bsCapabilitiesDefault = {
   acceptSslCerts: 'true',
   forceLocal: 'true',
   'browserstack.debug': 'true',
-  project: 'JS SDK',
+  project: 'Web SDK',
   'browserstack.user': process.env.BROWSERSTACK_USERNAME,
   'browserstack.key': process.env.BROWSERSTACK_ACCESS_KEY,
+  'browserstack.selenium_version': '3.141.59',
+  'browserstack.sendKeys': true,
   'browserstack.local': 'true',
   'browserstack.ie.enablePopups': 'false',
   unexpectedAlertBehaviour: 'dismiss',
@@ -53,7 +55,6 @@ const chromeOptions = {
 // https://github.com/ringcentral/testring/pull/63/files
 chromeCapabilities.set('goog:chromeOptions', chromeOptions)
 
-/* eslint-disable indent */
 const createDriver = ({ name, localIdentifier }) => (browser) =>
   browser.remote
     ? new Builder()
@@ -69,19 +70,16 @@ const createDriver = ({ name, localIdentifier }) => (browser) =>
     : new Builder()
         .forBrowser(browser.browserName)
         .withCapabilities(chromeCapabilities)
-/* eslint-enable indent */
 
 const createBrowser = async (browser, testCase) => {
   const localIdentifier = random()
 
-  /* eslint-disable indent */
   const bsLocal = browser.remote
     ? await createBrowserStackLocal({
         ...browserstackLocalDefault,
         localIdentifier,
       })
     : null
-  /* eslint-enable indent */
 
   const driver = await createDriver({
     name: testCase.file,
@@ -94,7 +92,6 @@ const createBrowser = async (browser, testCase) => {
   })
 
   if (browser.remote) driver.setFileDetector(new remote.FileDetector())
-  /* eslint-disable indent */
   const quitAll = async () => {
     await Promise.all([
       driver.quit(),
@@ -107,7 +104,6 @@ const createBrowser = async (browser, testCase) => {
         console.log('error finishing browser', e)
       })
   }
-  /* eslint-enable indent */
 
   driver.finish = async () => {
     console.log('finishing browser')
@@ -149,9 +145,9 @@ const createMocha = (driver, testCase) => {
 
 const printTestInfo = (browser, testCase) => {
   console.log(
-    browser.device
-      ? `Running ${testCase.file} on ${browser.device}`
-      : `Running ${testCase.file} against ${browser.browserName} on ${browser.os}`
+    browser.os
+      ? `Running ${testCase.file} against ${browser.browserName} on ${browser.os}`
+      : `Running ${testCase.file} on ${browser.browserName}`
   )
 }
 
@@ -163,7 +159,6 @@ const runner = async () => {
       const currentBrowser = browser.browserName
       let driver
       try {
-        console.log('Browser:', currentBrowser)
         driver = await createBrowser(browser, testCase)
         const mocha = createMocha(driver, testCase)