+++ webpack.config.babel.js
@@ -181,7 +181,7 @@ const basePlugins = (bundle_name) => [
       // ref: https://en.wikipedia.org/wiki/Base32
       // NOTE: please leave the BASE_32_VERSION be! It is updated automatically by
       // the release script 🤖
-      BASE_32_VERSION: 'BM',
+      BASE_32_VERSION: 'BT',
       PRIVACY_FEATURE_ENABLED: false,
       JWT_FACTORY: CONFIG.JWT_FACTORY,
       US_JWT_FACTORY: CONFIG.US_JWT_FACTORY,
@@ -202,9 +202,8 @@ const baseConfig = {
     extensions: ['.jsx', '.js', '.scss', '.json'],
     modules: [`${__dirname}/node_modules`, `${__dirname}/src`],
     alias: {
-      react: 'preact-compat',
-      'react-dom': 'preact-compat',
-      'react-modal': 'react-modal-onfido',
+      react: 'preact/compat',
+      'react-dom': 'preact/compat',
       '~utils': `${__dirname}/src/components/utils`,
     },
   },
@@ -230,7 +229,7 @@ const baseConfig = {
     setImmediate: false,
   },
 
-  devtool: PRODUCTION_BUILD ? 'source-map' : undefined,
+  devtool: PRODUCTION_BUILD ? 'source-map' : 'eval-cheap-source-map',
 }
 
 const configDist = {
@@ -274,6 +273,16 @@ const configDist = {
                 output: {
                   preamble: `/* Onfido SDK ${packageJson.version} */`,
                   comments: '/^!/',
+                },
+              },
+              extractComments: {
+                condition: /^\**!|@preserve|@license|@cc_on/i,
+                filename: (filename) => {
+                  const filenameNoExtension = path.basename(filename, '.min.js')
+                  return `${filenameNoExtension}.LICENSES.txt`
+                },
+                banner: (licenseFile) => {
+                  return `License information can be found in ${licenseFile}`
                 },
               },
             }),